Fixed #31
There was a difference in the initial values of the v and t registers between the Debug and Release builds. I thought that since the initial values can't be found online that it would be ok to leave these uninitialized. However, since these registers are actually only 15-bits long, there was no possibility of the NES 0'ing a 1 at the MSB. I thought of just adding `& 0x7f` around each call of v and t but that seemed a bit over the top compared to the simple fix of initalizing them to 0.